+++ src/open-actions/index.ts
@@ -1,3 +1,2 @@
 export * from './OpenActionMetadataSchema.js';
 export * from './OpenActionSchemaId.js';
-export * from './EventParameterSchema.js';